Added structures

+++ include/math/COperators.hpp
@@ -1,11 +1,11 @@
-#pragma once 
+#pragma once
 #include <complex>
 
 namespace math{
 
     std::complex<double> operator*(std::complex<double> p1, int p2)
     {
-        std::complex<double> a(p1.real()*p2, p1.imag()*p2);
+        std::complex<double> a(p1.real()*(double)p2, p1.imag()*(double)p2);
         return a;
     }
 }
